tracker/requirements.txt
numpy
psycopg2-binary

#sudo apt-get install python3-pigpio

#sudo apt-get update && sudo apt-get install python3-rpi.gpio

